chore: decouple support settings
