+++ products/conversations/frontend/scenes/settings/SupportSettingsScene.tsx
@@ -1,6 +1,6 @@
 import { useActions, useValues } from 'kea'
 
-import { IconCopy, IconPencil, IconPlus, IconRefresh, IconTrash } from '@posthog/icons'
+import { IconPlus } from '@posthog/icons'
 import {
     LemonBanner,
     LemonButton,
@@ -10,15 +10,11 @@ import {
     LemonInput,
     LemonSelect,
     LemonSwitch,
-    LemonTag,
     Link,
 } from '@posthog/lemon-ui'
 
-import { CodeSnippet } from 'lib/components/CodeSnippet'
 import { MemberSelectMultiple } from 'lib/components/MemberSelectMultiple'
 import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
-import { LemonDialog } from 'lib/lemon-ui/LemonDialog'
-import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { SceneExport } from 'scenes/sceneTypes'
 import { teamLogic } from 'scenes/teamLogic'
 
@@ -28,537 +24,16 @@ import { SceneTitleSection } from '~/layout/scenes/components/SceneTitleSection'
 import { ProductKey } from '~/queries/schema/schema-general'
 
 import { ScenesTabs } from '../../components/ScenesTabs'
+import { AuthorizedDomains } from './AuthorizedDomains'
 import { BrowserNotificationsSection } from './BrowserNotificationsSection'
+import { EmailSection } from './EmailSection'
+import { SecretApiKeySection } from './SecretApiKeySection'
+import { SlackSection } from './SlackSection'
 import { supportSettingsLogic } from './supportSettingsLogic'
 
 export const scene: SceneExport = {
     component: SupportSettingsScene,
     productKey: ProductKey.CONVERSATIONS,
-}
-
-function AuthorizedDomains(): JSX.Element {
-    const { conversationsDomains, isAddingDomain, editingDomainIndex, domainInputValue } =
-        useValues(supportSettingsLogic)
-    const { setDomainInputValue, saveDomain, removeDomain, startEditDomain, cancelDomainEdit } =
-        useActions(supportSettingsLogic)
-
-    return (
-        <div className="flex flex-col gap-2">
-            {conversationsDomains.length === 0 && !isAddingDomain && (
-                <div className="border rounded p-4 text-secondary">
-                    <p className="mb-0">
-                        <span className="font-bold">No domains configured.</span>
-                        <br />
-                        The widget will show on all domains. Add domains to limit where it appears.
-                    </p>
-                </div>
-            )}
-
-            {(isAddingDomain || editingDomainIndex !== null) && (
-                <div className="border rounded p-2 bg-surface-primary">
-                    <div className="gap-2">
-                        <LemonInput
-                            autoFocus
-                            value={domainInputValue}
-                            onChange={setDomainInputValue}
-                            placeholder="https://example.com or https://*.example.com"
-                            fullWidth
-                            onKeyDown={(e) => {
-                                if (e.key === 'Enter') {
-                                    saveDomain(domainInputValue, editingDomainIndex)
-                                } else if (e.key === 'Escape') {
-                                    cancelDomainEdit()
-                                }
-                            }}
-                        />
-                        <div className="flex gap-2 mt-2">
-                            <LemonButton
-                                type="primary"
-                                size="small"
-                                onClick={() => saveDomain(domainInputValue, editingDomainIndex)}
-                                disabledReason={!domainInputValue.trim() ? 'Enter a domain' : undefined}
-                            >
-                                Save
-                            </LemonButton>
-                            <LemonButton type="secondary" size="small" onClick={cancelDomainEdit}>
-                                Cancel
-                            </LemonButton>
-                        </div>
-                    </div>
-                </div>
-            )}
-
-            {conversationsDomains.map((domain: string, index: number) =>
-                editingDomainIndex === index ? null : (
-                    <div key={index} className="border rounded flex items-center p-2 pl-4 bg-surface-primary">
-                        <span title={domain} className="flex-1 truncate">
-                            {domain}
-                        </span>
-                        <div className="flex gap-1 shrink-0">
-                            <LemonButton
-                                icon={<IconPencil />}
-                                onClick={() => startEditDomain(index)}
-                                tooltip="Edit"
-                                size="small"
-                            />
-                            <LemonButton
-                                icon={<IconTrash />}
-                                tooltip="Remove domain"
-                                size="small"
-                                onClick={() => {
-                                    LemonDialog.open({
-                                        title: <>Remove {domain}?</>,
-                                        description: 'Are you sure you want to remove this domain?',
-                                        primaryButton: {
-                                            status: 'danger',
-                                            children: 'Remove',
-                                            onClick: () => removeDomain(index),
-                                        },
-                                        secondaryButton: {
-                                            children: 'Cancel',
-                                        },
-                                    })
-                                }}
-                            />
-                        </div>
-                    </div>
-                )
-            )}
-        </div>
-    )
-}
-
-function EmailSection(): JSX.Element {
-    return (
-        <SceneSection
-            title="Email channel"
-            description="Receive customer emails as support tickets. Set up forwarding from your email provider to route emails into conversations."
-            className="mt-4"
-        >
-            <LemonCard hoverEffect={false} className="flex flex-col gap-y-2 max-w-[800px] px-4 py-3">
-                <EmailChannelSection />
-            </LemonCard>
-        </SceneSection>
-    )
-}
-
-function EmailChannelSection(): JSX.Element {
-    const { emailConnected, emailFromEmail, emailFromName, emailConnecting, emailForwardingAddress } =
-        useValues(supportSettingsLogic)
-    const { setEmailFromEmail, setEmailFromName, connectEmail, disconnectEmail } = useActions(supportSettingsLogic)
-
-    return (
-        <div className="flex flex-col gap-y-2">
-            {!emailConnected ? (
-                <>
-                    <div>
-                        <label className="font-medium">Connect email</label>
-                        <p className="text-xs text-muted-alt">
-                            Enter the email address customers will contact you at (e.g. support@company.com). We'll give
-                            you a forwarding address to set up in your email provider.
-                        </p>
-                    </div>
-                    <div className="flex flex-col gap-2">
-                        <LemonInput
-                            value={emailFromEmail}
-                            onChange={(value) => setEmailFromEmail(value)}
-                            placeholder="support@company.com"
-                            fullWidth
-                        />
-                        <LemonInput
-                            value={emailFromName}
-                            onChange={(value) => setEmailFromName(value)}
-                            placeholder="Display name (e.g. Acme Support)"
-                            fullWidth
-                        />
-                        <div>
-                            <LemonButton
-                                type="primary"
-                                size="small"
-                                onClick={connectEmail}
-                                loading={emailConnecting}
-                                disabledReason={
-                                    !emailFromEmail || !emailFromName
-                                        ? 'Enter email address and display name'
-                                        : undefined
-                                }
-                            >
-                                Connect email
-                            </LemonButton>
-                        </div>
-                    </div>
-                </>
-            ) : (
-                <>
-                    <div>
-                        <label className="font-medium">Connection</label>
-                        <LemonTag type="success" className="ml-2">
-                            Connected
-                        </LemonTag>
-                    </div>
-                    {emailForwardingAddress && (
-                        <>
-                            <LemonDivider />
-                            <div>
-                                <label className="font-medium">Forwarding address</label>
-                                <p className="text-xs text-muted-alt">
-                                    Set up a forwarding rule in your email provider to forward incoming emails to this
-                                    address.
-                                </p>
-                                <div className="flex items-center gap-2 mt-1">
-                                    <code className="bg-surface-primary px-2 py-1 rounded text-sm break-all">
-                                        {emailForwardingAddress}
-                                    </code>
-                                    <LemonButton
-                                        type="secondary"
-                                        size="small"
-                                        icon={<IconCopy />}
-                                        onClick={() => {
-                                            void navigator.clipboard.writeText(emailForwardingAddress)
-                                            lemonToast.success('Copied to clipboard')
-                                        }}
-                                    />
-                                </div>
-                            </div>
-                            <LemonDivider />
-                            <div>
-                                <label className="font-medium">Setup instructions</label>
-                                <div className="text-xs text-muted-alt mt-1 flex flex-col gap-1">
-                                    <p className="mb-0">
-                                        <strong>Gmail:</strong> Settings → Forwarding → Add a forwarding address → paste
-                                        the address above → confirm.
-                                    </p>
-                                    <p className="mb-0">
-                                        <strong>Outlook:</strong> Settings → Mail → Forwarding → Enable forwarding →
-                                        paste the address above.
-                                    </p>
-                                    <p className="mb-0">
-                                        <strong>Other:</strong> Set up a forwarding rule to send all incoming emails to
-                                        the address above.
-                                    </p>
-                                </div>
-                            </div>
-                        </>
-                    )}
-                    <LemonDivider />
-                    <div className="flex justify-end">
-                        <LemonButton
-                            type="secondary"
-                            status="danger"
-                            size="small"
-                            onClick={() => {
-                                LemonDialog.open({
-                                    title: 'Disconnect email?',
-                                    description:
-                                        'This will stop creating tickets from emails. Existing tickets will not be affected.',
-                                    primaryButton: {
-                                        status: 'danger',
-                                        children: 'Disconnect',
-                                        onClick: disconnectEmail,
-                                    },
-                                    secondaryButton: { children: 'Cancel' },
-                                })
-                            }}
-                        >
-                            Disconnect email
-                        </LemonButton>
-                    </div>
-                </>
-            )}
-        </div>
-    )
-}
-
-function SlackSection(): JSX.Element {
-    return (
-        <SceneSection
-            title="SupportHog Slack bot"
-            description={
-                <>
-                    Add the SupportHog bot to your Slack workspace to create and manage support tickets directly from
-                    Slack messages.{' '}
-                    <Link to="https://posthog.com/docs/support/slack" target="_blank">
-                        Docs
-                    </Link>
-                </>
-            }
-            className="mt-4"
-        >
-            <LemonCard hoverEffect={false} className="flex flex-col gap-y-2 max-w-[800px] px-4 py-3">
-                <SlackChannelSection />
-            </LemonCard>
-        </SceneSection>
-    )
-}
-
-function SlackChannelSection(): JSX.Element {
-    const {
-        slackConnected,
-        slackChannelId,
-        slackChannels,
-        slackChannelsLoading,
-        slackTicketEmoji,
-        slackTicketEmojiValue,
-        slackBotIconUrl,
-        slackBotIconUrlValue,
-        slackBotDisplayName,
-        slackBotDisplayNameValue,
-    } = useValues(supportSettingsLogic)
-    const {
-        connectSlack,
-        setSlackChannel,
-        loadSlackChannelsWithToken,
-        setSlackTicketEmojiValue,
-        saveSlackTicketEmoji,
-        setSlackBotIconUrlValue,
-        setSlackBotDisplayNameValue,
-        saveSlackBotSettings,
-        disconnectSlack,
-    } = useActions(supportSettingsLogic)
-
-    return (
-        <div className="flex flex-col gap-y-2">
-            <div>
-                <label className="font-medium">Connection</label>
-                <p className="text-xs text-muted-alt">
-                    Install the SupportHog bot in your Slack workspace to enable support ticket creation from channels,
-                    mentions, and emoji reactions. This is separate from the main PostHog Slack integration.
-                </p>
-                {!slackConnected && (
-                    <LemonButton
-                        className="mt-2"
-                        type="primary"
-                        size="small"
-                        onClick={() => connectSlack(window.location.pathname)}
-                    >
-                        Add SupportHog to Slack
-                    </LemonButton>
-                )}
-            </div>
-            {slackConnected && (
-                <>
-                    <LemonDivider />
-                    <div className="gap-4">
-                        <div>
-                            <label className="font-medium">Support channel</label>
-                            <p className="text-xs text-muted-alt">
-                                Messages posted in this channel will automatically create support tickets. Thread
-                                replies become ticket messages.
-                            </p>
-                        </div>
-                        <div className="flex gap-2 items-center">
-                            <LemonSelect
-                                value={slackChannelId}
-                                options={[
-                                    { value: null, label: 'None' },
-                                    ...slackChannels.map((c: { id: string; name: string }) => ({
-                                        value: c.id,
-                                        label: `#${c.name}`,
-                                    })),
-                                ]}
-                                onChange={(value) => {
-                                    const channel = slackChannels.find((c: { id: string }) => c.id === value)
-                                    setSlackChannel(value, channel?.name ?? null)
-                                }}
-                                loading={slackChannelsLoading}
-                                placeholder="Select channel"
-                            />
-                            <LemonButton
-                                type="secondary"
-                                size="small"
-                                onClick={loadSlackChannelsWithToken}
-                                disabledReason={slackChannelsLoading ? 'Loading channels...' : undefined}
-                            >
-                                Refresh
-                            </LemonButton>
-                        </div>
-                    </div>
-                    <LemonDivider />
-                    <div className="flex items-center gap-4 justify-between">
-                        <div>
-                            <label className="font-medium">Ticket emoji trigger</label>
-                            <p className="text-xs text-muted-alt">
-                                React with this emoji on any message to create a support ticket from it.
-                            </p>
-                        </div>
-                        <div className="flex gap-2 items-center">
-                            <LemonInput
-                                value={slackTicketEmojiValue ?? slackTicketEmoji}
-                                onChange={setSlackTicketEmojiValue}
-                                placeholder="ticket"
-                                className="max-w-[200px]"
-                            />
-                            <LemonButton
-                                type="primary"
-                                size="small"
-                                onClick={saveSlackTicketEmoji}
-                                disabledReason={!slackTicketEmojiValue ? 'Enter an emoji name' : undefined}
-                            >
-                                Save
-                            </LemonButton>
-                        </div>
-                    </div>
-                    <LemonDivider />
-                    <div className="flex flex-col gap-2">
-                        <div>
-                            <label className="font-medium">Bot appearance</label>
-                            <p className="text-xs text-muted-alt">
-                                Override the bot's display name and icon when posting messages. Leave blank to use
-                                defaults. Requires the bot to be re-authorized if it was connected before this feature
-                                was available.
-                            </p>
-                        </div>
-                        <div className="flex items-center gap-2">
-                            <LemonInput
-                                value={slackBotDisplayNameValue ?? slackBotDisplayName ?? ''}
-                                onChange={setSlackBotDisplayNameValue}
-                                placeholder="Display name (e.g. SupportHog)"
-                                className="flex-1"
-                            />
-                        </div>
-                        <div className="flex items-center gap-2">
-                            <LemonInput
-                                value={slackBotIconUrlValue ?? slackBotIconUrl ?? ''}
-                                onChange={setSlackBotIconUrlValue}
-                                placeholder="Icon URL (e.g. https://example.com/icon.png)"
-                                className="flex-1"
-                            />
-                        </div>
-                        <div>
-                            <LemonButton
-                                type="primary"
-                                size="small"
-                                onClick={saveSlackBotSettings}
-                                disabledReason={
-                                    slackBotDisplayNameValue === null && slackBotIconUrlValue === null
-                                        ? 'No changes to save'
-                                        : undefined
-                                }
-                            >
-                                Save
-                            </LemonButton>
-                        </div>
-                    </div>
-                    <LemonDivider />
-                    <div className="flex items-center gap-4 justify-between">
-                        <div>
-                            <label className="font-medium">Bot mention</label>
-                            <p className="text-xs text-muted-alt">
-                                Users can @mention the bot in any channel to create a support ticket.
-                            </p>
-                        </div>
-                        <LemonTag type="success">Active</LemonTag>
-                    </div>
-                    <LemonDivider />
-                    <div className="flex justify-end">
-                        <LemonButton
-                            type="secondary"
-                            status="danger"
-                            size="small"
-                            onClick={() => {
-                                LemonDialog.open({
-                                    title: 'Remove SupportHog bot?',
-                                    description:
-                                        'This will stop creating tickets from Slack messages. Existing tickets will not be affected.',
-                                    primaryButton: {
-                                        status: 'danger',
-                                        children: 'Remove',
-                                        onClick: disconnectSlack,
-                                    },
-                                    secondaryButton: { children: 'Cancel' },
-                                })
-                            }}
-                        >
-                            Remove SupportHog bot
-                        </LemonButton>
-                    </div>
-                </>
-            )}
-        </div>
-    )
-}
-
-function SecretApiKeySection(): JSX.Element {
-    const { currentTeam, isTeamTokenResetAvailable } = useValues(teamLogic)
-    const { rotateSecretToken, deleteSecretTokenBackup } = useActions(teamLogic)
-
-    const openRotateDialog = (): void => {
-        const verb = currentTeam?.secret_api_token ? 'Rotate' : 'Generate'
-        const description = currentTeam?.secret_api_token
-            ? 'This will generate a new secret API key and move the existing one to backup. The old key will remain active until you delete it.'
-            : 'This will generate a new secret API key for authenticating external API requests.'
-
-        LemonDialog.open({
-            title: `${verb} secret API key?`,
-            description,
-            primaryButton: {
-                children: verb,
-                onClick: rotateSecretToken,
-            },
-            secondaryButton: { children: 'Cancel' },
-        })
-    }
-
-    return (
-        <SceneSection
-            title="Secret API key"
-            description="Used to authenticate external API requests (e.g. from workflows). This is the same key shown in Feature flags settings."
-            className="mt-4"
-        >
-            <LemonCard hoverEffect={false} className="flex flex-col gap-y-2 max-w-[800px] px-4 py-3">
-                <div>
-                    <h3 className="text-sm font-semibold mb-1">
-                        Primary key{' '}
-                        {currentTeam?.secret_api_token && <span className="text-green-700 text-xs ml-2">(Active)</span>}
-                    </h3>
-                    <CodeSnippet
-                        actions={
-                            <LemonButton
-                                icon={<IconRefresh />}
-                                size="xsmall"
-                                onClick={openRotateDialog}
-                                disabledReason={
-                                    !isTeamTokenResetAvailable
-                                        ? 'You do not have permission to rotate this key'
-                                        : undefined
-                                }
-                                tooltip={currentTeam?.secret_api_token ? 'Rotate key' : 'Generate key'}
-                            />
-                        }
-                        className={currentTeam?.secret_api_token ? '' : 'text-muted'}
-                        thing="Secret API key"
-                    >
-                        {currentTeam?.secret_api_token || 'Click the generate button on the right to create a new key.'}
-                    </CodeSnippet>
-                </div>
-
-                {currentTeam?.secret_api_token_backup ? (
-                    <div>
-                        <h3 className="text-sm font-semibold mb-1">
-                            Backup key <span className="text-orange-600 text-xs ml-2">(Pending deletion)</span>
-                        </h3>
-                        <CodeSnippet
-                            actions={
-                                <LemonButton
-                                    icon={<IconTrash />}
-                                    size="xsmall"
-                                    onClick={() => deleteSecretTokenBackup()}
-                                    tooltip="Delete backup key"
-                                />
-                            }
-                            thing="Backup secret API key"
-                        >
-                            {currentTeam.secret_api_token_backup}
-                        </CodeSnippet>
-                        <p className="text-xs text-muted mt-1">
-                            This key is still active to support services using the previous key. Delete it once you've
-                            fully migrated.
-                        </p>
-                    </div>
-                ) : null}
-            </LemonCard>
-        </SceneSection>
-    )
 }
 
 export function SupportSettingsScene(): JSX.Element {
@@ -980,7 +455,6 @@ export function SupportSettingsScene(): JSX.Element {
                             )}
                         </LemonCard>
                     </SceneSection>
-                    <SecretApiKeySection />
                     <SceneSection
                         title="Workflows"
                         description={
@@ -1074,6 +548,7 @@ export function SupportSettingsScene(): JSX.Element {
                             </div>
                         </LemonCard>
                     </SceneSection>
+                    <SecretApiKeySection />
                 </>
             )}
         </SceneContent>
